Extract renderer tools from component tools
