+++ projects/ngx-testing-tools/src/lib/common/tools/renderer/action/utils/output.ts
