+++ projects/ngx-testing-tools/src/lib/common/tools/renderer/query/models/component-query-tools.model.ts
