+++ projects/ngx-testing-tools/src/lib/v3-removed/index.ts
@@ -2,15 +2,15 @@ export { expectModuleToCreate } from './export-module-to-create';
 export { testPipeValues } from './test-pipe-values';
 export { fromInjector } from './from-injector';
 
-export { findComponent } from '../component-test-bed/tools/query/utils/find-component';
-export { findAllComponents } from '../component-test-bed/tools/query/utils/find-all-components';
-export { findElement } from '../component-test-bed/tools/query/utils/find-element';
-export { findAllElements } from '../component-test-bed/tools/query/utils/find-all-elements';
-export { findDebugElement } from '../component-test-bed/tools/query/utils/find-debug-element';
-export { findAllDebugElements } from '../component-test-bed/tools/query/utils/find-all-debug-elements';
+export { findComponent } from '../common/tools/renderer/query/utils/find-component';
+export { findAllComponents } from '../common/tools/renderer/query/utils/find-all-components';
+export { findElement } from '../common/tools/renderer/query/utils/find-element';
+export { findAllElements } from '../common/tools/renderer/query/utils/find-all-elements';
+export { findDebugElement } from '../common/tools/renderer/query/utils/find-debug-element';
+export { findAllDebugElements } from '../common/tools/renderer/query/utils/find-all-debug-elements';
 
-export { click } from '../component-test-bed/tools/action/utils/click';
-export { emitOutput } from '../component-test-bed/tools/action/utils/output';
+export { click } from '../common/tools/renderer/action/utils/click';
+export { emitOutput } from '../common/tools/renderer/action/utils/output';
 
 export { emitFakeSuccessResponse } from '../common/tools/http/utils/emit-success-response';
 export { emitFakeErrorResponse } from '../common/tools/http/utils/emit-error-response';
