+++ projects/ngx-testing-tools/src/tests/units/component-test-bed/query/cannot-find.spec.ts
@@ -1,4 +1,4 @@
-import { throwCannotFind } from '../../../../lib/component-test-bed/tools/query/utils/cannot-find';
+import { throwCannotFind } from '../../../../lib/common/tools/renderer/query/utils/cannot-find';
 import { NoWhereDirective } from '../../../fixtures/directives/no-where.directive';
 
 describe('throwCannotFind', () => {
